refactor: 사용자 입력 유효성 체크 부분 구조 개선
